+++ src/main/java/vendingmachine/domain/coin/Change.java
@@ -1,8 +1,10 @@
-package vendingmachine.domain;
+package vendingmachine.domain.coin;
 
 import java.util.EnumMap;
 import java.util.Map;
 import java.util.Map.Entry;
+import vendingmachine.domain.Coin;
+import vendingmachine.domain.VendingMachine;
 
 public class Change {
